+++ src/common/hooks/index.js
@@ -1,3 +1,4 @@
 export { default as useCloseReasonOptions } from './useCloseReasonOptions';
 export * from './useLinesLimit';
 export * from './useVendor';
+export * from './useAcqMethods';