+++ src/settings/OrderTemplates/OrderTemplatesEditor/POLineDetailsForm/POLineDetailsForm.test.js
@@ -1,12 +1,23 @@
 import React from 'react';
+import { QueryClient, QueryClientProvider } from 'react-query';
 import { render, screen } from '@testing-library/react';
 import { Form } from 'react-final-form';
 
 import POLineDetailsForm from './POLineDetailsForm';
 
+const queryClient = new QueryClient();
+
+// eslint-disable-next-line react/prop-types
+const wrapper = ({ children }) => (
+  <QueryClientProvider client={queryClient}>
+    {children}
+  </QueryClientProvider>
+);
+
 const defaultProps = {
   formValues: {},
   createInventorySetting: {},
+  acquisitionMethods: [],
 };
 
 const renderPOLineDetailsForm = (props = {}) => render(
@@ -19,6 +30,7 @@ const renderPOLineDetailsForm = (props = {}) => render(
       />
     )}
   />,
+  { wrapper },
 );
 
 describe('POLineDetailsForm', () => {