+++ test/bigtest/network/configs/index.js
@@ -2,3 +2,4 @@ export * from './configClosingReasons';
 export * from './configPrefixes';
 export * from './configSuffixes';
 export * from './configTemplates';
+export * from './configAcqMethods';